Initial commit of my work so far.
Everything is in the James folder for now. The make-*.py files write files
that describe a collection of spheres ray-tracing2.py is the current approach
to drawing the shadows of these objects. Just got this working today.

ray-tracing2.py doesn't really do ray tracing. It expects command line arguments
including an infile (generated by one of the make-*.py scripts) and an outfile
(something.jpeg where it will save the image).