test1.py uses the ADParser.py module to read art description files.
The region of the plane to render was changed in make-art-description.py to demonstrate that this change is respected in ray-tracing2.py and also test1.py.

A bug that caused an IndexError in ray-tracing2.py in rare circumstances was fixed. This code might still be broken, but I'm planning to change the whole way rendering is done anyway.